pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "code-review-bot"
version = "2.0.0"
description = "AI-powered code review bot with CLI and Web Dashboard"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "core"},
    {include = "providers"},
    {include = "models"},
    {include = "cli"},
    {include = "web"}
]
homepage = "https://github.com/yourusername/review-bot"
repository = "https://github.com/yourusername/review-bot"
documentation = "https://github.com/yourusername/review-bot#readme"
keywords = ["code-review", "ai", "git", "claude", "chatgpt", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

[tool.poetry.dependencies]
python = "^3.11"
# Core dependencies
pydantic = "^2.5.0"
gitpython = "^3.1.40"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"

# CLI dependencies
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
questionary = "^2.0.1"

# AI providers
anthropic = "^0.25.0"
openai = "^1.12.0"
google-generativeai = "^0.4.0"

# Web dashboard
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
websockets = "^12.0"
aiofiles = "^23.2.1"

# HTTP client
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
review-bot = "cli.main:app"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["core", "providers", "models", "cli", "web"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=core --cov=providers --cov=models --cov=cli --cov=web --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"